update init table scripts

+++ src/main/resources/db/init/create_score_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE Score
 (
     id          INT AUTO_INCREMENT PRIMARY KEY,
-    user_id   INT  NOT NULL,
+    user_id   BIGINT  NOT NULL,
     score_value INT  NOT NULL,
     created_at  TIMESTAMP NOT NULL,
     FOREIGN KEY (user_id) REFERENCES User (id)
